index, list, add, update de company y contract -- revisar parte de enumeraciones

+++ src/main/java/seniorsathome/seniorsathomespring/dao/ContractRowMapper.java
@@ -18,7 +18,7 @@ public class ContractRowMapper implements RowMapper<Contract> {
         contract.setQuantity(rs.getInt("quantity"));
         contract.setStartDate(rs.getDate("startDate"));
         contract.setFinalDate(rs.getDate("finalDate"));
-        contract.setPrice(rs.getInt("price"));
+        contract.setPrice(rs.getDouble("price"));
         contract.setCompanyID(rs.getString("companyID"));
 
         return contract;
